Added complete code to the ProductList class and also created an AppException class.
The AppException class has the meaning of storing all exceptions that will be used in our app. One exception used in ProductList is now stored in there as the first exception. ProductList is now complete with connection with the firebase database and can add, remove, and search from database, as well as from its local copy (local list).